+++ README.md
@@ -1,12 +1,9 @@
-# ANSIcht
+# ansidote
 
 <img alt="Screenshot" src="docs/screen.webp"/>
 
-(A play on the words ANSI and "Ansicht",
-the german word for "point of view" or "perspective".)
-
 # Description
-ANSIcht is a simple text art editor that supports
+ansidote is a simple text art editor that supports
 24bit colors and unicode text. Right now, the feature
 set is very limited to a basic, on-the-fly calculated
 color palette and pre-defined text characters, as
@@ -19,22 +16,20 @@ Currently, ANSIcht exports and reads from .txt files
 containing characters and ANSI escape sequences.
 
 # Installation
-    git clone https://github.com/dewberryants/ANSIcht.git
-    cd ANSIcht
+    git clone https://github.com/dewberryants/ansidote.git
+    cd ansidote
     pip install . --user
 
 # Use
-    ansicht
+    ansidote
 
 # Known Issues
- * Palette Scaling is very jumpy
  * No scrolling on palette / character map
- * No custom palette
  * Dark Mode Only
 
 # Closing Words
 
 Many thanks go out to the developers of pygame,
-which ANSIcht is mostly developed in.
+which ansidote is mostly developed in.
 
 Happy painting!
