+++ ansidote/resources/settings.bin
